Convert ipc/ to TypeScript
Now that main/ is converted to TypeScript, ipc/ can also be converted.
This removes the need to maintain separate declaration files.
