+++ src/ipc/error.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-export const cleanIpcErrorMessage = (message, replacecment = '') =>
+export const cleanIpcErrorMessage = (message: string, replacecment = '') =>
     message.replace(
         /Error invoking remote method '.*?': *(Error:)? */,
         replacecment
